Integrate router-flux
