+++ app/App.js
@@ -5,8 +5,11 @@ import {
   Text,
 } from 'react-native';
 
+import { Router, Scene } from 'react-native-router-flux';
 import { Provider } from 'react-redux';
-import configureStore from './configureStore'
+import configureStore from './configureStore';
+
+import Intro from './Intro';
 
 const store = configureStore();
 
@@ -14,38 +17,11 @@ export default class App extends Component {
   render() {
     return (
       <Provider store={store}>
-        <View style={styles.container}>
-          <Text style={styles.welcome}>
-            Welcome to React Native!
-          </Text>
-          <Text style={styles.instructions}>
-            To get started, edit index.ios.js
-          </Text>
-          <Text style={styles.instructions}>
-            Press Cmd+R to reload,{'\n'}
-            Cmd+D or shake for dev menu
-          </Text>
-        </View>      
+        <Router key='root'>
+          <Scene key='Intro' component={Intro} initial={true}/>
+          <Scene key='Intro2' component={Intro}/>
+        </Router>
       </Provider>
     );
   }
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#F5FCFF',
-  },
-  welcome: {
-    fontSize: 20,
-    textAlign: 'center',
-    margin: 10,
-  },
-  instructions: {
-    textAlign: 'center',
-    color: '#333333',
-    marginBottom: 5,
-  },
-});
